working around reducers
