+++ src/pages/index.js
@@ -2,7 +2,7 @@ import React from "react"
 import { Link } from "gatsby"
 import App from "../components/App.js"
 import { Provider } from "react-redux"
-import store from "./store"
+import store from "../reducers/books"
 
 const IndexPage = () => (
   <div>
